Controller operations  refactored.
Some handlers were need to Include(r=>r.ForeignKeyModel) before sending
it to keep it simple in frontend side. So it is added to the related
api controllers. Secondly, the delete operation of some models need to
be controlled if there is foreign key between any table that causes
conflict. So if a Fixture is getting removed from system, also the
fixture result object of this fixture is deleted.